demo/requirements.txt
pytorch
numpy
python_speech_feature
scipy